Added ref argument to cc_sea(), source load_maps.R
`cc_sea()` requires an internet connection to pull down the map of the
oceans every time it is called. I downloaded the default source map so we
only need to do this once. This is a slight speed up, but more importantly
means we aren't hooped when the internet goes down.

Also source the new `load_maps.R` file, where the seaRef map is downloaded
or read from file.