This updated gap analysis now runs in less than 30 secs on my PC. A vast improvement from the >3 hrs the previous analysis took.
To help with debugging and ease of intpertation i created seperate functions for the srs, ers and grs scores.

I am still tweaking this analysis. I want to restrict the area of suitable habitat considered to be within the area of historical suitability, so even if area of high suitability expands in the future into new PAs, those areas will be excluded.